+++ src/Oro/Bundle/ApiBundle/Tests/Unit/Processor/Subresource/ChangeRelationship/LoadNormalizedRelationshipTest.php
@@ -100,7 +100,7 @@ class LoadNormalizedRelationshipTest extends ChangeRelationshipProcessorTestCase
         $this->context->setParentId(123);
         $this->context->setParentEntity(new \stdClass());
         $this->context->setIsCollection($isCollection);
-        $this->context->setMasterRequest(true);
+        $this->context->setMainRequest(true);
         $this->context->setCorsRequest(true);
         $this->context->setHateoas(true);
         $this->context->getRequestHeaders()->set('test-header', 'some value');
@@ -109,7 +109,7 @@ class LoadNormalizedRelationshipTest extends ChangeRelationshipProcessorTestCase
         $expectedGetContext = new GetContext($this->configProvider, $this->metadataProvider);
         $expectedGetContext->setVersion($this->context->getVersion());
         $expectedGetContext->getRequestType()->set($this->context->getRequestType());
-        $expectedGetContext->setMasterRequest(false);
+        $expectedGetContext->setMainRequest(false);
         $expectedGetContext->setCorsRequest(false);
         $expectedGetContext->setHateoas(true);
         $expectedGetContext->setRequestHeaders($this->context->getRequestHeaders());
@@ -152,7 +152,7 @@ class LoadNormalizedRelationshipTest extends ChangeRelationshipProcessorTestCase
         $expectedContext->setAction(ApiAction::UPDATE_RELATIONSHIP);
         $expectedContext->setVersion($this->context->getVersion());
         $expectedContext->getRequestType()->set($this->context->getRequestType());
-        $expectedContext->setMasterRequest(true);
+        $expectedContext->setMainRequest(true);
         $expectedContext->setCorsRequest(true);
         $expectedContext->setHateoas(true);
         $expectedContext->setRequestHeaders($this->context->getRequestHeaders());