+++ src/Oro/Bundle/NavigationBundle/Menu/Matcher/Voter/RoutePatternVoter.php
@@ -22,9 +22,8 @@ class RoutePatternVoter implements VoterInterface
     #[\Override]
     public function matchItem(ItemInterface $item): ?bool
     {
-        // Using master request, as sub-requests routes must not be taken into account when matching the menu items
+        // Use main request, as sub-requests routes must not be taken into account when matching the menu items
         $request = $this->requestStack->getMainRequest();
-
         if (null === $request) {
             return null;
         }